bookgrid_proj/requirements.txt
asgiref==3.3.1
boto3==1.17.105
botocore==1.20.105
certifi==2021.5.30
chardet==4.0.0
coreapi==2.3.3
coreschema==0.0.4
dj-database-url==0.5.0
dj-email-url==1.0.2
Django==3.1.7
django-cache-url==3.2.3
django-cors-headers==3.7.0
django-heroku==0.3.1
django-rest-framework==0.1.0
django-rest-passwordreset==1.2.0
django-rest-swagger==2.2.0
django-storages==1.11.1
djangorestframework==3.12.2
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.6.0
environs==8.0.0
gunicorn==19.9.0
idna==2.10
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.1
jmespath==0.10.0
MarkupSafe==2.0.1
marshmallow==3.12.1
openapi-codec==1.3.2
packaging==20.9
Pillow==8.2.0
psycopg2==2.8.6
psycopg2-binary==2.8.5
PyJWT==1.7.1
pyparsing==2.4.7
python-dateutil==2.8.1
python-dotenv==0.18.0
pytz==2021.1
PyYAML==5.1.2
requests==2.25.1
ruamel.yaml==0.17.10
ruamel.yaml.clib==0.2.2
s3transfer==0.4.2
simplejson==3.17.2
six==1.16.0
sqlparse==0.4.1
uritemplate==3.0.1
urllib3==1.26.5
whitenoise==5.2.0
